IG-1327 refactored in line with other entity add/edit forms
